Add files via upload
The "new frame.FCStd" is the one you want to open.
I know that the sizing is different from Manolis his current model, but that is because it fits the sizing of parts that are locally available. We can easily change it. I will start adding the parts of Manolis his design over the weekend.